+++ src/SignalR/common/Shared/testCertECC.pfx
